pythonanywhere_deploy/requirements.txt
# IA Poste Manager - Production Requirements

# Core Framework (FastAPI)
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
asgiref==3.7.2

# Legacy Flask (à supprimer progressivement)
flask==3.0.0
flask-cors==4.0.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# ORM & Migrations
prisma==0.11.0
alembic==1.13.1

# Authentication & Security
pyjwt==2.8.0
cryptography==41.0.7
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utilities
requests==2.31.0
python-dateutil==2.8.2
aiohttp==3.9.1
httpx==0.26.0

# Logging
colorlog==6.8.0

# AI & ML
openai==1.10.0

# Testing
pytest==8.3.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Monitoring
prometheus-client==0.19.0

# Production WSGI Server (pour compatibilité)
gunicorn==21.2.0
waitress==2.1.2